small adjustments, move test to user folder
