+++ user/user.go
@@ -1,7 +1,6 @@
 package user
 
 import (
-	"database/sql"
 	"encoding/json"
 	"net/http"
 	"timezone-converter/db"
@@ -37,6 +36,7 @@ func Register(w http.ResponseWriter, r *http.Request) {
 	u.Timeslots = defaultTimeslots
 	jsonWithTimeslots, _ := json.Marshal(defaultTimeslots)
 	json.NewDecoder(r.Body).Decode(&u)
+	// TODO: probably need to replace any with something
 	args := []any{u.Username, u.Password, jsonWithTimeslots}
 	db.Db_instance.Exec("INSERT INTO users(username, password, timeslots) values(?,?,?)", args...)
 
@@ -46,8 +46,7 @@ func Register(w http.ResponseWriter, r *http.Request) {
 func BookTime(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	userId := params["userId"]
-	db, _ := sql.Open("sqlite3", "./users.db")
-	row := db.QueryRow("SELECT * FROM users WHERE id=?", userId)
+	row := db.Db_instance.QueryRow("SELECT * FROM users WHERE id=?", userId)
 	user := User{}
 	row.Scan(&user.Id, &user.Username, &user.Password, &user.Timeslots)
